patra_agent/edge_server/requirements.txt
blinker==1.8.2
click==8.1.7
filelock==3.16.1
Flask==3.0.3
fsspec==2024.9.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
pillow==10.4.0
sympy==1.13.3
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.12.2
Werkzeug==3.0.4
zipp==3.20.2
requests==2.26.0
python-dotenv~=1.0.1
kafka-python==2.0.2
confluent-kafka
huggingface_hub